feat(806): add API to filter tasks by status
