+++ server/src/routes/course/tasks/index.ts
@@ -1,0 +1,4 @@
+export { getCourseTask } from './getCourseTask';
+export { getCourseTasks } from './getCourseTasks';
+export { createCourseTaskDistribution } from './createCourseTaskDistribution';
+export { getCourseTasksDetails } from './getCourseTasksDetails';
